change naming convention of types and models
Currently both, the model and the type, have the name of the entity. Still
both might be used at once, which makes naming them differently necessary.
Also a file should be names after the default export. in the future Types
will be named with the entity's name followed by "Type" and the model is
named after with the entity's name followed by "Model".